Refactor useDocument to use the new API
