+++ src/renderer/backend/documents/useDocument.ts
@@ -1,21 +1,16 @@
 import { useState, useEffect } from "react";
 
-type PropsType = {
-  courseId: string;
-  documentId: string;
-};
-
-const useDocument = ({ courseId, documentId }: PropsType) => {
+const useDocument = (documentId: string) => {
   const [document, setDocument] = useState<Doc | null>(null);
 
   useEffect(() => {
     const fetchDocument = async () => {
-      const doc = await window.api.document.get(courseId, documentId);
+      const doc = await window.api.document.get(documentId);
       setDocument(doc);
     };
 
     fetchDocument();
-  }, [courseId, documentId]);
+  }, [documentId]);
 
   return document;
 };
